Adding conctact works. Working on fixing showing correct names in Private Chat

+++ MessageStack/MessageStack/Controllers/ContactController.cs
@@ -3,31 +3,43 @@ using MessageStack.Repositories;
 using System;
 using System.Linq;
 using System.Web.Mvc;
+using MessageStack.Context;
 
 namespace MessageStack.Controllers
 {
-    [Authorize]
-    public class ContactController : Controller
+    public class ContactController : BaseController
     {
-        private readonly ContactRepository _contactRepository;
-        private readonly AccountRepository _accountRepository;
 
-        public ContactController() { }
-
-        public ContactController(ContactRepository contactRepository, AccountRepository accountRepository)
+        public ContactController() : this (RepositoryContext.GetInstance())
         {
-            _contactRepository = contactRepository;
-            _accountRepository = accountRepository;
+
+        }
+
+        public ContactController(MessageStackContext messageStackContext) : base(messageStackContext)
+        {
+
         }
 
         public ActionResult Index()
         {
-            return View(_contactRepository.GetWhere(c => c.OwnerAccount == Session["Loggedin_Account"]).Result.ToList());
+            if (!IsLoggedIn()) return RedirectToAction("Index", "Home");
+            var loggedInAccount = (Account) Session["Loggedin_Account"];
+            var contacts = _contactRepository.GetWhere(c => c.OwnerAccount.Id == loggedInAccount.Id).ToList();
+
+            foreach (var contact in contacts)
+            {
+                contact.OwnerAccount = loggedInAccount;
+                contact.TargetAccount = _accountRepository.FirstOrDefault(a => a.Phonenumber == contact.Phonenumber);
+            }
+
+            return View(contacts);
         }
 
         [HttpPost]
         public ActionResult RemoveContact(Guid id)
         {
+            if (!IsLoggedIn()) return RedirectToAction("Index", "Home");
+
             if (ModelState.IsValid)
             {
                 var result = _contactRepository.Remove(id);
@@ -40,24 +52,28 @@ namespace MessageStack.Controllers
 
         public ActionResult CreateContact()
         {
+            if (!IsLoggedIn()) return RedirectToAction("Index", "Home");
+
             return View();
         }
 
         [HttpPost]
         public ActionResult CreateContact(Contact contact)
         {
+            if (!IsLoggedIn()) return RedirectToAction("Index", "Home");
+
             if (ModelState.IsValid)
             {
                 if (Session?["Loggedin_Account"] == null) return View(contact);
 
-                var account = _accountRepository.FirstOrDefault(a => a.Phonenumber == contact.Phonenumber).Result;
+                var account = _accountRepository.FirstOrDefault(a => a.Phonenumber == contact.Phonenumber);
 
                 if (account.Email == contact.Email)
                 {
                     contact.OwnerAccount = (Account)Session["Loggedin_Account"];
                     contact.TargetAccount = account;
 
-                    var result = _contactRepository.Add(contact).Result;
+                    var result = _contactRepository.Add(contact);
                     return View("~/Views/Contact/Index.cshtml");
                 }
 
@@ -71,12 +87,20 @@ namespace MessageStack.Controllers
             return View(contact);
         }
 
-        public ActionResult ChangeContact(Guid id) => View(_contactRepository.GetById(id).Result);
+        public ActionResult ChangeContact(Guid id)
+        {
+            if (!IsLoggedIn()) return RedirectToAction("Index", "Home");
+
+            return View(_contactRepository.GetById(id));
+        }
+
 
 
         [HttpPost]
         public ActionResult ChangeContact(Contact contact)
         {
+            if (!IsLoggedIn()) return RedirectToAction("Index", "Home");
+
             var result = _contactRepository.Update(contact);
             return RedirectToAction("Index");
         }
